# optimization: Reduce cache lookups in CCoinsViewCache::FetchCoin