@paplorinc 2024-06-25
Thanks for the compiler hints @fanquake, I had a lot of problems with MacOSX14.sdk, so was using `Apple clang version 15.0.0 (clang-1500.3.9.4)`. Might give GCC and libstdc++ another try if I can find a readme of how to do that on a mac properly, everything just failed catastrophically so far.

<details>
<summary>
And thanks for the IBD hint @sipa, by limiting it to height 200k and disabling unrelated validation the target method was indeed the main bottleneck (~20% of time spent):
<img src="https://github.com/bitcoin/bitcoin/assets/1841944/e391aa54-9f09-4361-a339-0e34d9212928">
</summary>
```patch
Index: src/validation.cpp
IDEA additional info:
Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
<+>UTF-8
===================================================================
diff --git a/src/validation.cpp b/src/validation.cpp
--- a/src/validation.cpp	(revision 538363738e9e30813cf3e76ca4f71c1aaff349e7)
+++ b/src/validation.cpp	(date 1719299940880)
@@ -2133,7 +2133,7 @@
                        bool cacheFullScriptStore, PrecomputedTransactionData& txdata,
                        std::vector<CScriptCheck>* pvChecks)
 {
-    if (tx.IsCoinBase()) return true;
+    return true;
 
     if (pvChecks) {
         pvChecks->reserve(tx.vin.size());
@@ -2969,6 +2969,10 @@
         coins_tip.DynamicMemoryUsage() * (1.0 / (1 << 20)),
         coins_tip.GetCacheSize(),
         !warning_messages.empty() ? strprintf(" warning='%s'", warning_messages) : "");
+
+    if (tip->nHeight == 200'000) {
+        throw std::runtime_error{"Enough!"};
+    }
 }
</details>

I ran `bitcoind -reindex-chainstate -assumevalid=0 -par=1` for 20 iterations (skipped the first in both as a warmup) as such:
```bash
git checkout master && make -j10 >/dev/null 2>&1 && for i in {1..20}; do (time (./src/bitcoind -reindex-chainstate -assumevalid=0 -par=1 > /dev/null 2>&1)) 2>&1 | awk '{print $(NF-1)}'; done; git checkout paplorinc/FetchCoin && make -j10 >/dev/null 2>&1 && for i in {1..20}; do (time (./src/bitcoind -reindex-chainstate -assumevalid=0 -par=1 > /dev/null 2>&1)) 2>&1 | awk '{print $(NF-1)}'; done
```

which reveals a small but measurable difference between the two runs (~4%):
<img src="https://github.com/bitcoin/bitcoin/assets/1841944/c0a6e6b2-f2a3-4edb-9f4c-dc73b33af95b">

Is there any other test or benchmark or change you'd like me to do?